Move canonical version number to pyproject.toml
AutoPub expects the canonical version number to be stored in the
pyproject.toml file with a key of `version` and will fail in its
absence. This moves the canonical location of the version number to
pyproject.toml, but since AutoPub will automatically increment the
version number in `crispy_tailwind/__init__.py`, this remain DRY and
without any need to manually manage version numbers in multiple places.